-add: big news image
The big image will now now the newsfeed's article image and adds a clickable title to open the article in default webbrowser.
Downloaded images are cached in the data folder.
Moved the default news item to a resource to be used when newsfeed image fails to load.
A little code cleanup.
Added possible fix for crashes at character select screen (option_sound)