-add: Auto-Launch option
Added a autoLaunch option to the .aelcf setting files.
When set to true, the launcher will start, check if the specified server is online, and then automatically press the play button (make sure username and password are included)
Then 10 seconds after launching the game, it will close the launcher.
This is useful for local testing if you need multiple accounts. Just create a separate config file with username and password included, and launch by clicking the settings file.
Also added more defaults for the settings file class itself.