-add: Document Folder detection
Added code to try and detect where the client's document folder will be.
This is done so updates to system.cfg can be done correctly for modded or otherwise non-standard clients.
This is done using a pre-defined string search within the exe-file. Not the most ideal way of doing this as it doesn't work for encrypted executables. (will throw a warning)